Unravel contracts: separate message sending functions
