+++ src/test/beacon_tests.cpp
@@ -4,6 +4,7 @@
 #include "beacon.h"
 #include "appcache.h"
 #include "contract/contract.h"
+#include "contract/message.h"
 
 #include <boost/test/unit_test.hpp>
 #include <boost/algorithm/hex.hpp>
